Update nginx.conf to set read timeout via environment variable or default to 60s (which is nginx default)
Rename nginx.conf to nginx.conf.template as we will be replacing the environment variable with a value.

Update Dockerfile to install gettext (for envsubst) and to copy the nginx.conf.template to /app instead of copying the old nginx.conf.

Update run_and_monitor.sh to use envsubst to set the nginx read timeout from environment variable.